bin/requirements.txt
absl-py==2.3.1
asttokens==3.0.0
astunparse==1.6.3
bokeh==3.4.3
Bottleneck==1.4.2
Brotli==1.0.9
brotlicffi==1.0.9.2
cached-property==1.5.2
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.3.2
cicflowmeter==0.1.6
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.2.1
cupy==13.6.0
cycler==0.11.0
cytoolz==1.0.1
dask==2024.5.0
dask-expr==1.1.0
debugpy==1.8.16
decorator==5.2.1
distributed==2024.5.0
exceptiongroup==1.3.0
executing==2.2.0
fastrlock==0.8.3
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2025.7.0
gast==0.6.0
gmpy2==2.2.1
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.14.0
HeapDict==1.0.1
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.0
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.4
llvmlite==0.43.0
lmdb==1.6.2
locket==1.0.0
lz4==4.3.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
ml-dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
namex==0.1.0
nest_asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numexpr==2.10.1
numpy==1.26.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
parso==0.8.4
partd==1.4.2
pickleshare==0.7.5
pillow==11.1.0
pip==25.3
platformdirs==4.3.8
polars==1.34.0
polars-runtime-32==1.34.0
prompt_toolkit==3.0.51
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.23
pyflowmeter==0.2.4
Pygments==2.19.2
pyod==1.1.3
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.2
requests==2.32.5
rich==14.1.0
scapy==2.5.0
scikit-learn==1.6.1
scikit-multiflow==0.5.3
scipy==1.11.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
stack_data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tblib==3.1.0
tensorboard==2.19.0
tensorboard_data_server==0.7.0
tensorflow_cpu==2.19.1
termcolor==3.1.0
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
unicodedata2==15.1.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
win-inet-pton==1.1.0
wrapt==1.17.3
xgboost==2.1.3
xyzservices==2022.9.0
zict==3.0.0
zipp==3.21.0
